Add docker login support to the base utility for Docker commands
This change means that all LAVA actions that rely on this utility to
interact with docker can add support for `docker login` by
initializing the utility object appropriately. Actions must separately
enable the schema support for the login stanza.

This requires `docker_secure` to be set if docker login is to be
permitted.

Signed-off-by: Ed Smith <ed.smith@collabora.com>